mod to IPV4Address to use uint32_t instead of uint16_t to address wrap-around issues
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@7225 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ MOOS_V10.0.3_May1215/MOOSCore/Core/libMOOS/Utils/IPV4Address.cpp
@@ -95,7 +95,7 @@ void IPV4Address::set_host(const std::string & host)
 	host_=host;
 }
 
-void IPV4Address::set_port(uint16_t port)
+void IPV4Address::set_port(uint32_t port)
 {
 	port_=port;
 }
@@ -105,7 +105,7 @@ std::string IPV4Address::host() const
 	return host_;
 }
 
-uint16_t IPV4Address::port() const
+uint32_t IPV4Address::port() const
 {
 	return port_;
 }
